+++ src/_pytask/resolve_dependencies.py
@@ -1,20 +1,29 @@
-import pprint
+import itertools
 import sys
-import traceback
+from typing import Dict
+from typing import List
+from typing import Tuple
 
-import click
 import networkx as nx
 from _pytask.config import hookimpl
+from _pytask.console import ARROW_DOWN_ICON
+from _pytask.console import console
+from _pytask.console import FILE_ICON
+from _pytask.console import TASK_ICON
 from _pytask.dag import node_and_neighbors
 from _pytask.dag import task_and_descending_tasks
 from _pytask.dag import TopologicalSorter
 from _pytask.database import State
+from _pytask.enums import ColorCode
 from _pytask.exceptions import NodeNotFoundError
 from _pytask.exceptions import ResolvingDependenciesError
 from _pytask.mark import Mark
 from _pytask.nodes import reduce_node_name
+from _pytask.path import find_common_ancestor_of_nodes
 from _pytask.report import ResolvingDependenciesReport
 from pony import orm
+from rich.traceback import Traceback
+from rich.tree import Tree
 
 
 @hookimpl
@@ -86,11 +95,11 @@ def pytask_resolve_dependencies_select_execution_dag(dag):
 
 
 @hookimpl
-def pytask_resolve_dependencies_validate_dag(session, dag):
+def pytask_resolve_dependencies_validate_dag(dag):
     """Validate the DAG."""
     _check_if_dag_has_cycles(dag)
-    _check_if_root_nodes_are_available(dag, session)
-    _check_if_tasks_have_the_same_products(dag, session)
+    _check_if_root_nodes_are_available(dag)
+    _check_if_tasks_have_the_same_products(dag)
 
 
 def _have_task_or_neighbors_changed(task_name, dag):
@@ -126,15 +135,28 @@ def _check_if_dag_has_cycles(dag):
     else:
         raise ResolvingDependenciesError(
             "The DAG contains cycles which means a dependency is directly or "
-            "implicitly a product of the same task. See the following tuples "
-            "(from a to b) to see the path in the graph which defines the cycle."
-            f"\n\n{pprint.pformat(cycles)}"
+            "indirectly a product of the same task. See the following the path of "
+            "nodes in the graph which forms the cycle."
+            f"\n\n{_format_cycles(cycles)}"
         )
 
 
-def _check_if_root_nodes_are_available(dag, session):
-    paths = session.config["paths"]
-    missing_root_nodes = {}
+def _format_cycles(cycles: List[Tuple[str]]) -> str:
+    """Format cycles as a paths connected by arrows."""
+    chain = [x for i, x in enumerate(itertools.chain(*cycles)) if i % 2 == 0]
+    chain += [cycles[-1][1]]
+
+    lines = chain[:1]
+    for x in chain[1:]:
+        lines.append("     " + ARROW_DOWN_ICON)
+        lines.append(x)
+    text = "\n".join(lines)
+
+    return text
+
+
+def _check_if_root_nodes_are_available(dag):
+    missing_root_nodes = []
 
     for node in dag.nodes:
         is_node = "node" in dag.nodes[node]
@@ -144,56 +166,97 @@ def _check_if_root_nodes_are_available(dag, session):
                 dag.nodes[node]["node"].state()
             except NodeNotFoundError:
                 # Shorten node names for better printing.
-                short_node_name = reduce_node_name(dag.nodes[node]["node"], paths)
-                short_successors = _reduce_names_of_multiple_nodes(
-                    dag.successors(node), dag, paths
-                )
-                missing_root_nodes[short_node_name] = short_successors
+                missing_root_nodes.append(node)
 
     if missing_root_nodes:
+        all_names = missing_root_nodes + [
+            successor
+            for node in missing_root_nodes
+            for successor in dag.successors(node)
+        ]
+        common_ancestor = find_common_ancestor_of_nodes(*all_names)
+        dictionary = {}
+        for node in missing_root_nodes:
+            short_node_name = reduce_node_name(
+                dag.nodes[node]["node"], [common_ancestor]
+            )
+            short_successors = _reduce_names_of_multiple_nodes(
+                dag.successors(node), dag, [common_ancestor]
+            )
+            dictionary[short_node_name] = short_successors
+
+        text = _format_dictionary_to_tree(dictionary, "Missing dependencies:")
         raise ResolvingDependenciesError(
-            "There are some dependencies missing which do not exist and are not "
-            "produced by any task. See the following dictionary with missing nodes as "
-            "keys and dependent tasks as values."
-            f"\n\n{pprint.pformat(missing_root_nodes)}"
+            "Some dependencies do not exist or are not produced by any task. See the "
+            "following tree which shows which dependencies are missing for which tasks."
+            f"\n\n{text}"
         )
 
 
-def _check_if_tasks_have_the_same_products(dag, session):
-    paths = session.config["paths"]
-    nodes_created_by_multiple_tasks = {}
+def _format_dictionary_to_tree(dict_: Dict[str, List[str]], title: str) -> str:
+    """Format missing root nodes."""
+    tree = Tree(title)
+
+    for node, tasks in dict_.items():
+        branch = tree.add(FILE_ICON + node)
+        for task in tasks:
+            branch.add(TASK_ICON + task)
+
+    text = "".join(
+        [x.text for x in tree.__rich_console__(console, console.options)][:-1]
+    )
+
+    return text
+
+
+def _check_if_tasks_have_the_same_products(dag):
+    nodes_created_by_multiple_tasks = []
 
     for node in dag.nodes:
         is_node = "node" in dag.nodes[node]
         if is_node:
             parents = list(dag.predecessors(node))
             if len(parents) > 1:
-                # Reduce node names for better printing.
-                short_node = reduce_node_name(dag.nodes[node]["node"], paths)
-                short_parents = _reduce_names_of_multiple_nodes(parents, dag, paths)
-                nodes_created_by_multiple_tasks[short_node] = short_parents
+                nodes_created_by_multiple_tasks.append(node)
 
     if nodes_created_by_multiple_tasks:
+        all_names = nodes_created_by_multiple_tasks + [
+            predecessor
+            for node in nodes_created_by_multiple_tasks
+            for predecessor in dag.predecessors(node)
+        ]
+        common_ancestor = find_common_ancestor_of_nodes(*all_names)
+        dictionary = {}
+        for node in nodes_created_by_multiple_tasks:
+            short_node_name = reduce_node_name(
+                dag.nodes[node]["node"], [common_ancestor]
+            )
+            short_predecessors = _reduce_names_of_multiple_nodes(
+                dag.predecessors(node), dag, [common_ancestor]
+            )
+            dictionary[short_node_name] = short_predecessors
+        text = _format_dictionary_to_tree(dictionary, "Products from multiple tasks:")
         raise ResolvingDependenciesError(
             "There are some tasks which produce the same output. See the following "
-            "dictionary with products as keys and their producing tasks as values."
-            f"\n\n{pprint.pformat(nodes_created_by_multiple_tasks)}"
+            "tree which shows which products are produced by multiple tasks."
+            f"\n\n{text}"
         )
 
 
 @hookimpl
-def pytask_resolve_dependencies_log(session, report):
+def pytask_resolve_dependencies_log(report):
     """Log errors which happened while resolving dependencies."""
-    tm_width = session.config["terminal_width"]
+    console.print()
+    console.rule(
+        f"[{ColorCode.FAILED}]Failures during resolving dependencies",
+        style=ColorCode.FAILED,
+    )
 
-    click.echo(f"{{:=^{tm_width}}}".format(" Failures during resolving dependencies "))
+    console.print()
+    console.print(Traceback.from_exception(*report.exc_info))
 
-    click.echo("")
-
-    traceback.print_exception(*report.exc_info)
-
-    click.echo("")
-    click.echo("=" * tm_width)
+    console.print()
+    console.rule(style=ColorCode.FAILED)
 
 
 def _reduce_names_of_multiple_nodes(names, dag, paths):