+++ tests/test_collect_command.py
@@ -1,7 +1,6 @@
 import textwrap
 
 import pytest
-from _pytask.collect_command import _organize_tasks
 from _pytask.collect_command import _print_collected_tasks
 from pytask import cli
 
@@ -24,19 +23,23 @@ def test_collect_task(runner, tmp_path):
 
     result = runner.invoke(cli, ["collect", tmp_path.as_posix()])
 
-    assert "<Module" in result.output
-    assert "task_dummy.py>" in result.output
-    assert "<Function task_dummy>" in result.output
+    captured = result.output.replace("\n", "").replace(" ", "")
+    assert "<Module" in captured
+    assert "task_dummy.py>" in captured
+    assert "<Function" in captured
+    assert "task_dummy>" in captured
 
     result = runner.invoke(cli, ["collect", tmp_path.as_posix(), "--nodes"])
 
-    assert "<Module" in result.output
-    assert "task_dummy.py>" in result.output
-    assert "<Function task_dummy>" in result.output
-    assert "<Dependency" in result.output
-    assert "in.txt>" in result.output
-    assert "<Product" in result.output
-    assert "out.txt>" in result.output
+    captured = result.output.replace("\n", "").replace(" ", "")
+    assert "<Module" in captured
+    assert "task_dummy.py>" in captured
+    assert "<Function" in captured
+    assert "task_dummy>" in captured
+    assert "<Dependency" in captured
+    assert "in.txt>" in captured
+    assert "<Product" in captured
+    assert "out.txt>" in captured
 
 
 @pytest.mark.end_to_end
@@ -58,20 +61,25 @@ def test_collect_task_with_expressions(runner, tmp_path):
 
     result = runner.invoke(cli, ["collect", tmp_path.as_posix(), "-k", "_1"])
 
-    assert "<Module" in result.output
-    assert "task_dummy.py>" in result.output
-    assert "<Function task_dummy_1>" in result.output
-    assert "<Function task_dummy_2>" not in result.output
+    captured = result.output.replace("\n", "").replace(" ", "")
+    assert "<Module" in captured
+    assert "task_dummy.py>" in captured
+    assert "<Function" in captured
+    assert "task_dummy_1>" in captured
+    assert "<Function" in captured
+    assert "task_dummy_2>" not in captured
 
     result = runner.invoke(cli, ["collect", tmp_path.as_posix(), "-k", "_1", "--nodes"])
 
-    assert "<Module" in result.output
-    assert "task_dummy.py>" in result.output
-    assert "<Function task_dummy_1>" in result.output
-    assert "<Dependency" in result.output
-    assert "in_1.txt>" in result.output
-    assert "<Product" in result.output
-    assert "out_1.txt>" in result.output
+    captured = result.output.replace("\n", "").replace(" ", "")
+    assert "<Module" in captured
+    assert "task_dummy.py>" in captured
+    assert "<Function" in captured
+    assert "task_dummy_1>" in captured
+    assert "<Dependency" in captured
+    assert "in_1.txt>" in captured
+    assert "<Product" in captured
+    assert "out_1.txt>" in captured
 
 
 @pytest.mark.end_to_end
@@ -102,22 +110,27 @@ def test_collect_task_with_marker(runner, tmp_path, config_name):
 
     result = runner.invoke(cli, ["collect", tmp_path.as_posix(), "-m", "wip"])
 
-    assert "<Module" in result.output
-    assert "task_dummy.py>" in result.output
-    assert "<Function task_dummy_1>" in result.output
-    assert "<Function task_dummy_2>" not in result.output
+    captured = result.output.replace("\n", "").replace(" ", "")
+    assert "<Module" in captured
+    assert "task_dummy.py>" in captured
+    assert "<Function" in captured
+    assert "task_dummy_1>" in captured
+    assert "<Function" in captured
+    assert "task_dummy_2>" not in captured
 
     result = runner.invoke(
         cli, ["collect", tmp_path.as_posix(), "-m", "wip", "--nodes"]
     )
 
-    assert "<Module" in result.output
-    assert "task_dummy.py>" in result.output
-    assert "<Function task_dummy_1>" in result.output
-    assert "<Dependency" in result.output
-    assert "in_1.txt>" in result.output
-    assert "<Product" in result.output
-    assert "out_1.txt>" in result.output
+    captured = result.output.replace("\n", "").replace(" ", "")
+    assert "<Module" in captured
+    assert "task_dummy.py>" in captured
+    assert "<Function" in captured
+    assert "task_dummy_1>" in captured
+    assert "<Dependency" in captured
+    assert "in_1.txt>" in captured
+    assert "<Product" in captured
+    assert "out_1.txt>" in captured
 
 
 @pytest.mark.end_to_end
@@ -150,22 +163,27 @@ def test_collect_task_with_ignore_from_config(runner, tmp_path, config_name):
 
     result = runner.invoke(cli, ["collect", tmp_path.as_posix()])
 
-    assert "<Module" in result.output
-    assert "task_dummy_1.py>" in result.output
-    assert "task_dummy_2.py>" not in result.output
-    assert "<Function task_dummy_1>" in result.output
-    assert "<Function task_dummy_2>" not in result.output
+    captured = result.output.replace("\n", "").replace(" ", "")
+    assert "<Module" in captured
+    assert "task_dummy_1.py>" in captured
+    assert "task_dummy_2.py>" not in captured
+    assert "<Function" in captured
+    assert "task_dummy_1>" in captured
+    assert "<Function" in captured
+    assert "task_dummy_2>" not in captured
 
     result = runner.invoke(cli, ["collect", tmp_path.as_posix(), "--nodes"])
 
-    assert "<Module" in result.output
-    assert "task_dummy_1.py>" in result.output
-    assert "task_dummy_2.py>" not in result.output
-    assert "<Function task_dummy_1>" in result.output
-    assert "<Dependency" in result.output
-    assert "in_1.txt>" in result.output
-    assert "<Product" in result.output
-    assert "out_1.txt>" in result.output
+    captured = result.output.replace("\n", "").replace(" ", "")
+    assert "<Module" in captured
+    assert "task_dummy_1.py>" in captured
+    assert "task_dummy_2.py>" not in captured
+    assert "<Function" in captured
+    assert "task_dummy_1>" in captured
+    assert "<Dependency" in captured
+    assert "in_1.txt>" in captured
+    assert "<Product" in captured
+    assert "out_1.txt>" in captured
 
 
 @pytest.mark.end_to_end
@@ -193,43 +211,29 @@ def test_collect_task_with_ignore_from_cli(runner, tmp_path):
         cli, ["collect", tmp_path.as_posix(), "--ignore", "task_dummy_2.py"]
     )
 
-    assert "<Module" in result.output
-    assert "task_dummy_1.py>" in result.output
-    assert "task_dummy_2.py>" not in result.output
-    assert "<Function task_dummy_1>" in result.output
-    assert "<Function task_dummy_2>" not in result.output
+    captured = result.output.replace("\n", "").replace(" ", "")
+    assert "<Module" in captured
+    assert "task_dummy_1.py>" in captured
+    assert "task_dummy_2.py>" not in captured
+    assert "<Function" in captured
+    assert "task_dummy_1>" in captured
+    assert "<Function" in captured
+    assert "task_dummy_2>" not in captured
 
     result = runner.invoke(
         cli, ["collect", tmp_path.as_posix(), "--ignore", "task_dummy_2.py", "--nodes"]
     )
 
-    assert "<Module" in result.output
-    assert "task_dummy_1.py>" in result.output
-    assert "task_dummy_2.py>" not in result.output
-    assert "<Function task_dummy_1>" in result.output
-    assert "<Dependency" in result.output
-    assert "in_1.txt>" in result.output
-    assert "<Product" in result.output
-    assert "out_1.txt>" in result.output
-
-
-@pytest.mark.unit
-def test_organize_tasks():
-    dependency = DummyClass()
-    dependency.name = "dependency"
-
-    task = DummyClass()
-    task.name = "prefix::task_dummy"
-    task.path = "task_path.py"
-    task.depends_on = {0: dependency}
-    task.produces = {}
-
-    result = _organize_tasks([task])
-
-    assert "task_path.py" in result
-    assert "task_dummy" in result["task_path.py"]
-    assert "dependency" in result["task_path.py"]["task_dummy"]["depends_on"]
-    assert not result["task_path.py"]["task_dummy"]["produces"]
+    captured = result.output.replace("\n", "").replace(" ", "")
+    assert "<Module" in captured
+    assert "task_dummy_1.py>" in captured
+    assert "task_dummy_2.py>" not in captured
+    assert "<Function" in captured
+    assert "task_dummy_1>" in captured
+    assert "<Dependency" in captured
+    assert "in_1.txt>" in captured
+    assert "<Product" in captured
+    assert "out_1.txt>" in captured
 
 
 @pytest.mark.unit
@@ -243,7 +247,8 @@ def test_print_collected_tasks_without_nodes(capsys):
     captured = capsys.readouterr().out
 
     assert "<Module path.py>" in captured
-    assert "<Function task_dummy>" in captured
+    assert "<Function" in captured
+    assert "task_dummy>" in captured
     assert "<Dependency in.txt>" not in captured
     assert "<Product out.txt>" not in captured
 
@@ -259,6 +264,7 @@ def test_print_collected_tasks_with_nodes(capsys):
     captured = capsys.readouterr().out
 
     assert "<Module path.py>" in captured
-    assert "<Function task_dummy>" in captured
+    assert "<Function" in captured
+    assert "task_dummy>" in captured
     assert "<Dependency in.txt>" in captured
     assert "<Product out.txt>" in captured