+++ tests/test_resolve_dependencies.py
@@ -7,81 +7,72 @@ import networkx as nx
 import pytest
 from _pytask.exceptions import NodeNotFoundError
 from _pytask.exceptions import ResolvingDependenciesError
-from _pytask.nodes import MetaNode
-from _pytask.nodes import MetaTask
+from _pytask.nodes import FilePathNode
+from _pytask.nodes import PythonFunctionTask
 from _pytask.resolve_dependencies import _check_if_root_nodes_are_available
 from _pytask.resolve_dependencies import pytask_resolve_dependencies_create_dag
 from pytask import cli
 
 
 @attr.s
-class Task(MetaTask):
-    name = attr.ib(type=str)
-    depends_on = attr.ib(factory=dict)
-    produces = attr.ib(factory=dict)
-
-    def execute(self):
-        pass
-
-    def state(self):
-        pass
-
-
-@attr.s
-class Node(MetaNode):
+class Node(FilePathNode):
     """See https://github.com/python-attrs/attrs/issues/293 for property hack."""
 
     name = attr.ib(type=str)
+    value = attr.ib()
+    path = attr.ib()
 
     def state(self):
-        if self.name == "missing":
+        if "missing" in self.name:
             raise NodeNotFoundError
-
-
-class DummySession:
-    pass
 
 
 @pytest.mark.unit
 def test_create_dag():
-    task = Task(
-        name="task",
-        depends_on={0: Node(name="node_1"), 1: Node(name="node_2")},
+    root = Path.cwd() / "src"
+    task = PythonFunctionTask(
+        "task_dummy",
+        root.as_posix() + "::task_dummy",
+        root,
+        None,
+        depends_on={
+            0: Node.from_path(root / "node_1"),
+            1: Node.from_path(root / "node_2"),
+        },
     )
 
     dag = pytask_resolve_dependencies_create_dag([task])
 
-    assert sorted(dag.nodes) == ["node_1", "node_2", "task"]
+    assert all(
+        any(i in node for i in ["node_1", "node_2", "task"]) for node in dag.nodes
+    )
 
 
 @pytest.mark.unit
 def test_check_if_root_nodes_are_available():
     dag = nx.DiGraph()
 
-    root = Path("directory")
-    session = DummySession()
-    session.config = {"paths": [root]}
+    root = Path.cwd() / "src"
 
-    task = Task("task")
-    task.path = root.joinpath("task_dummy")
+    path = root.joinpath("task_dummy")
+    task = PythonFunctionTask("task", path.as_posix() + "::task", path, None)
+    task.path = path
     task.base_name = "task_dummy"
     dag.add_node(task.name, task=task)
 
-    available_node = Node("available")
-    available_node.path = root.joinpath("available_node")
+    available_node = Node.from_path(root.joinpath("available_node"))
     dag.add_node(available_node.name, node=available_node)
     dag.add_edge(available_node.name, task.name)
 
     with does_not_raise():
-        _check_if_root_nodes_are_available(dag, session)
+        _check_if_root_nodes_are_available(dag)
 
-    missing_node = Node("missing")
-    missing_node.path = root.joinpath("missing_node")
+    missing_node = Node.from_path(root.joinpath("missing_node"))
     dag.add_node(missing_node.name, node=missing_node)
     dag.add_edge(missing_node.name, task.name)
 
     with pytest.raises(ResolvingDependenciesError):
-        _check_if_root_nodes_are_available(dag, session)
+        _check_if_root_nodes_are_available(dag)
 
 
 @pytest.mark.end_to_end
@@ -103,7 +94,7 @@ def test_check_if_root_nodes_are_available_end_to_end(tmp_path, runner):
 
     # Ensure that node names are reduced.
     assert "Failures during resolving dependencies" in result.output
-    assert "There are some dependencies missing which do not" in result.output
+    assert "Some dependencies do not exist or are" in result.output
     assert tmp_path.joinpath("task_d.py").as_posix() + "::task_d" not in result.output
     assert tmp_path.name + "/task_d.py::task_d" in result.output
     assert tmp_path.joinpath("in.txt").as_posix() not in result.output
@@ -133,7 +124,7 @@ def test_check_if_root_nodes_are_available_with_separate_build_folder_end_to_end
 
     # Ensure that node names are reduced.
     assert "Failures during resolving dependencies" in result.output
-    assert "There are some dependencies missing which do not" in result.output
+    assert "Some dependencies do not exist" in result.output
     assert tmp_path.joinpath("task_d.py").as_posix() + "::task_d" not in result.output
     assert "src/task_d.py::task_d" in result.output
     assert tmp_path.joinpath("bld", "in.txt").as_posix() not in result.output